requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
build==1.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
ecdsa==0.19.1
fastapi==0.129.0
frozenlist==1.8.0
fsspec==2025.12.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
logzero==1.7.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.1
nautilus_trader==1.222.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pluggy==1.6.0
portion==2.6.1
prometheus_client==0.24.1
propcache==0.4.1
psutil==7.2.2
psycopg2-binary==2.9.11
pyarrow==23.0.1
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
Pygments==2.19.2
pyotp==2.9.0
pyproject_hooks==1.2.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
pytz==2025.2
PyYAML==6.0.3
redis==7.2.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
smartapi-python==1.5.5
sortedcontainers==2.4.0
starlette==0.52.1
TA-Lib==0.6.8
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
yarl==1.22.0
